+++ src/test/kotlin/me/bossm0n5t3r/leetcode/generateParentheses/GenerateParenthesesTest.kt
@@ -1,11 +1,11 @@
-package generateParentheses
+package me.bossm0n5t3r.leetcode.generateParentheses
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
 class GenerateParenthesesTest {
-    private val generateParentheses = GenerateParentheses()
+    private val generateParentheses = GenerateParentheses.Solution()
 
     data class GenerateParenthesesTestData(
         val n: Int,
